// aah_apr_front_end-master/types/Types.ts
export type TrainingBenefeciaryForm = {
  name: string;
  fatherHusbandName: string;
  gender: string;
  age: number;
  phone: string;
  code: string;
  email: string;
  participantOrganization: string;
  jobTitle: string;
  dateOfRegistration: string;
};